src/test/java/seedu/application/model/job/AddressTest.java
package seedu.application.model.job;

import org.junit.jupiter.api.Test;

import static org.junit.jupiter.api.Assertions.assertEquals;
import static org.junit.jupiter.api.Assertions.assertFalse;
import static org.junit.jupiter.api.Assertions.assertNotEquals;
import static org.junit.jupiter.api.Assertions.assertTrue;
import static seedu.application.testutil.Assert.assertThrows;

public class AddressTest {
    @Test
    public void constructor_null_throwsNullPointerException() {
        assertThrows(NullPointerException.class, () -> new Address(null));
    }

    @Test
    public void constructor_invalidAddress_throwsIllegalArgumentException() {
        String invalidAddress = "";
        assertThrows(IllegalArgumentException.class, () -> new Address(invalidAddress));
    }

    @Test
    void isValidAddress() {
        // null address
        assertThrows(NullPointerException.class, () -> Address.isValidAddress(null));

        // invalid addresses
        assertFalse(Address.isValidAddress("")); // empty string
        assertFalse(Address.isValidAddress(" ")); // spaces only
        assertFalse(Address.isValidAddress(" 123")); // starts with a white space

        // valid addresses
        assertTrue(Address.isValidAddress("Blk 456, Den Road, #01-355"));
        assertTrue(Address.isValidAddress("-")); // one character
        assertTrue(Address.isValidAddress("Leng Inc; 1234 Market St; San Francisco CA 2349879; USA")); // long address
    }

    @Test
    void testEquals() {
        Address address = new Address("Blk 456, Den Road, #01-355");

        // same values -> returns true
        assertEquals(address, new Address("Blk 456, Den Road, #01-355"));

        // same object -> returns true
        assertEquals(address, address);

        // null -> returns false
        assertNotEquals(address, null);

        // different types -> returns false
        assertNotEquals(address, new Company("Google"));

        // different values -> returns false
        assertNotEquals(address, new Address("Blk 123, Den Road, #01-355"));
    }

}

